Rename some methods to be more consistent
