+++ pkg/api/http.go
@@ -22,7 +22,6 @@ import (
 	"github.com/grafana/tempo/pkg/traceql"
 	"github.com/grafana/tempo/pkg/util"
 	"github.com/grafana/tempo/tempodb"
-	"github.com/grafana/tempo/tempodb/backend"
 )
 
 const (
@@ -253,343 +252,6 @@ func ParseSearchRequest(r *http.Request) (*tempopb.SearchRequest, error) {
 	if req.End <= req.Start {
 		return nil, fmt.Errorf("http parameter start must be before end. received start=%d end=%d", req.Start, req.End)
 	}
-	return req, nil
-}
-
-// ParseSearchBlockRequest parses all http parameters necessary to perform a block search.
-func ParseSearchBlockRequest(r *http.Request) (*tempopb.SearchBlockRequest, error) {
-	searchReq, err := ParseSearchRequest(r)
-	if err != nil {
-		return nil, err
-	}
-
-	// start and end = 0 is NOT fine for a block search request
-	if searchReq.End == 0 {
-		return nil, errors.New("start and end required")
-	}
-
-	req := &tempopb.SearchBlockRequest{
-		SearchReq: searchReq,
-	}
-
-	s := r.URL.Query().Get(urlParamStartPage)
-	startPage, err := strconv.ParseInt(s, 10, 32)
-	if err != nil {
-		return nil, fmt.Errorf("invalid startPage: %w", err)
-	}
-	if startPage < 0 {
-		return nil, fmt.Errorf("startPage must be non-negative. received: %s", s)
-	}
-	req.StartPage = uint32(startPage)
-
-	s = r.URL.Query().Get(urlParamPagesToSearch)
-	pagesToSearch64, err := strconv.ParseInt(s, 10, 32)
-	if err != nil {
-		return nil, fmt.Errorf("invalid pagesToSearch %s: %w", s, err)
-	}
-	if pagesToSearch64 <= 0 {
-		return nil, fmt.Errorf("pagesToSearch must be greater than 0. received: %s", s)
-	}
-	req.PagesToSearch = uint32(pagesToSearch64)
-
-	s = r.URL.Query().Get(urlParamBlockID)
-	blockID, err := uuid.Parse(s)
-	if err != nil {
-		return nil, fmt.Errorf("invalid blockID: %w", err)
-	}
-	req.BlockID = blockID.String()
-
-	s = r.URL.Query().Get(urlParamEncoding)
-	encoding, err := backend.ParseEncoding(s)
-	if err != nil {
-		return nil, err
-	}
-	req.Encoding = encoding.String()
-
-	s = r.URL.Query().Get(urlParamIndexPageSize)
-	indexPageSize, err := strconv.ParseInt(s, 10, 32)
-	if err != nil {
-		return nil, fmt.Errorf("invalid indexPageSize %s: %w", s, err)
-	}
-	req.IndexPageSize = uint32(indexPageSize)
-
-	s = r.URL.Query().Get(urlParamTotalRecords)
-	totalRecords, err := strconv.ParseInt(s, 10, 32)
-	if err != nil {
-		return nil, fmt.Errorf("invalid totalRecords %s: %w", s, err)
-	}
-	if totalRecords <= 0 {
-		return nil, fmt.Errorf("totalRecords must be greater than 0. received %d", totalRecords)
-	}
-	req.TotalRecords = uint32(totalRecords)
-
-	// Data encoding can be blank for some block formats, therefore
-	// no validation on the param here.  Eventually we may be able
-	// to remove this parameter entirely.
-	dataEncoding := r.URL.Query().Get(urlParamDataEncoding)
-	req.DataEncoding = dataEncoding
-
-	version := r.URL.Query().Get(urlParamVersion)
-	if version == "" {
-		return nil, errors.New("version required")
-	}
-	req.Version = version
-
-	s = r.URL.Query().Get(urlParamSize)
-	size, err := strconv.ParseUint(s, 10, 64)
-	if err != nil {
-		return nil, fmt.Errorf("invalid size %s: %w", s, err)
-	}
-	req.Size_ = size
-
-	// Footer size can be 0 for some blocks, just ensure we
-	// get a valid integer.
-	f := r.URL.Query().Get(urlParamFooterSize)
-	footerSize, err := strconv.ParseUint(f, 10, 32)
-	if err != nil {
-		return nil, fmt.Errorf("invalid footerSize %s: %w", f, err)
-	}
-	req.FooterSize = uint32(footerSize)
-
-	s = r.URL.Query().Get(urlParamDedicatedColumns)
-	if s != "" {
-		var dedicatedColumns []*tempopb.DedicatedColumn
-		err = json.Unmarshal([]byte(s), &dedicatedColumns)
-		if err != nil {
-			return nil, fmt.Errorf("invalid dedicatedColumns '%s': %w", s, err)
-		}
-		req.DedicatedColumns = dedicatedColumns
-	}
-
-	return req, nil
-}
-
-func ParseSearchTagValuesBlockRequest(r *http.Request) (*tempopb.SearchTagValuesBlockRequest, error) {
-	return parseSearchTagValuesBlockRequest(r, false)
-}
-
-func ParseSearchTagValuesBlockRequestV2(r *http.Request) (*tempopb.SearchTagValuesBlockRequest, error) {
-	return parseSearchTagValuesBlockRequest(r, true)
-}
-
-func parseSearchTagValuesBlockRequest(r *http.Request, enforceTraceQL bool) (*tempopb.SearchTagValuesBlockRequest, error) {
-	var tagSearchReq *tempopb.SearchTagValuesRequest
-	var err error
-	if !enforceTraceQL {
-		tagSearchReq, err = ParseSearchTagValuesRequest(r)
-	} else {
-		tagSearchReq, err = ParseSearchTagValuesRequestV2(r)
-	}
-
-	if err != nil {
-		return nil, err
-	}
-
-	// start and end = 0 is NOT fine for a block search request
-	if tagSearchReq.End == 0 {
-		return nil, errors.New("start and end required")
-	}
-
-	req := &tempopb.SearchTagValuesBlockRequest{
-		SearchReq: tagSearchReq,
-	}
-
-	s := r.URL.Query().Get(urlParamStartPage)
-	startPage, err := strconv.ParseInt(s, 10, 32)
-	if err != nil {
-		return nil, fmt.Errorf("invalid startPage: %w", err)
-	}
-	if startPage < 0 {
-		return nil, fmt.Errorf("startPage must be non-negative. received: %s", s)
-	}
-	req.StartPage = uint32(startPage)
-
-	s = r.URL.Query().Get(urlParamPagesToSearch)
-	pagesToSearch64, err := strconv.ParseInt(s, 10, 32)
-	if err != nil {
-		return nil, fmt.Errorf("invalid pagesToSearch %s: %w", s, err)
-	}
-	if pagesToSearch64 <= 0 {
-		return nil, fmt.Errorf("pagesToSearch must be greater than 0. received: %s", s)
-	}
-	req.PagesToSearch = uint32(pagesToSearch64)
-
-	s = r.URL.Query().Get(urlParamBlockID)
-	blockID, err := uuid.Parse(s)
-	if err != nil {
-		return nil, fmt.Errorf("invalid blockID: %w", err)
-	}
-	req.BlockID = blockID.String()
-
-	s = r.URL.Query().Get(urlParamEncoding)
-	encoding, err := backend.ParseEncoding(s)
-	if err != nil {
-		return nil, err
-	}
-	req.Encoding = encoding.String()
-
-	s = r.URL.Query().Get(urlParamIndexPageSize)
-	indexPageSize, err := strconv.ParseInt(s, 10, 32)
-	if err != nil {
-		return nil, fmt.Errorf("invalid indexPageSize %s: %w", s, err)
-	}
-	req.IndexPageSize = uint32(indexPageSize)
-
-	s = r.URL.Query().Get(urlParamTotalRecords)
-	totalRecords, err := strconv.ParseInt(s, 10, 32)
-	if err != nil {
-		return nil, fmt.Errorf("invalid totalRecords %s: %w", s, err)
-	}
-	if totalRecords <= 0 {
-		return nil, fmt.Errorf("totalRecords must be greater than 0. received %d", totalRecords)
-	}
-	req.TotalRecords = uint32(totalRecords)
-
-	// Data encoding can be blank for some block formats, therefore
-	// no validation on the param here.  Eventually we may be able
-	// to remove this parameter entirely.
-	dataEncoding := r.URL.Query().Get(urlParamDataEncoding)
-	req.DataEncoding = dataEncoding
-
-	version := r.URL.Query().Get(urlParamVersion)
-	if version == "" {
-		return nil, errors.New("version required")
-	}
-	req.Version = version
-
-	s = r.URL.Query().Get(urlParamSize)
-	size, err := strconv.ParseUint(s, 10, 64)
-	if err != nil {
-		return nil, fmt.Errorf("invalid size %s: %w", s, err)
-	}
-	req.Size_ = size
-
-	// Footer size can be 0 for some blocks, just ensure we
-	// get a valid integer.
-	f := r.URL.Query().Get(urlParamFooterSize)
-	footerSize, err := strconv.ParseUint(f, 10, 32)
-	if err != nil {
-		return nil, fmt.Errorf("invalid footerSize %s: %w", f, err)
-	}
-	req.FooterSize = uint32(footerSize)
-
-	s = r.URL.Query().Get(urlParamDedicatedColumns)
-	if s != "" {
-		var dedicatedColumns []*tempopb.DedicatedColumn
-		err = json.Unmarshal([]byte(s), &dedicatedColumns)
-		if err != nil {
-			return nil, fmt.Errorf("invalid dedicatedColumns '%s': %w", s, err)
-		}
-		req.DedicatedColumns = dedicatedColumns
-	}
-
-	return req, nil
-}
-
-func ParseSearchTagsBlockRequest(r *http.Request) (*tempopb.SearchTagsBlockRequest, error) {
-	tagSearchReq, err := ParseSearchTagsRequest(r)
-	if err != nil {
-		return nil, err
-	}
-
-	// start and end = 0 is NOT fine for a block search request
-	if tagSearchReq.End == 0 {
-		return nil, errors.New("start and end required")
-	}
-
-	req := &tempopb.SearchTagsBlockRequest{
-		SearchReq: tagSearchReq,
-	}
-
-	s := r.URL.Query().Get(urlParamStartPage)
-	startPage, err := strconv.ParseInt(s, 10, 32)
-	if err != nil {
-		return nil, fmt.Errorf("invalid startPage: %w", err)
-	}
-	if startPage < 0 {
-		return nil, fmt.Errorf("startPage must be non-negative. received: %s", s)
-	}
-	req.StartPage = uint32(startPage)
-
-	s = r.URL.Query().Get(urlParamPagesToSearch)
-	pagesToSearch64, err := strconv.ParseInt(s, 10, 32)
-	if err != nil {
-		return nil, fmt.Errorf("invalid pagesToSearch %s: %w", s, err)
-	}
-	if pagesToSearch64 <= 0 {
-		return nil, fmt.Errorf("pagesToSearch must be greater than 0. received: %s", s)
-	}
-	req.PagesToSearch = uint32(pagesToSearch64)
-
-	s = r.URL.Query().Get(urlParamBlockID)
-	blockID, err := uuid.Parse(s)
-	if err != nil {
-		return nil, fmt.Errorf("invalid blockID: %w", err)
-	}
-	req.BlockID = blockID.String()
-
-	s = r.URL.Query().Get(urlParamEncoding)
-	encoding, err := backend.ParseEncoding(s)
-	if err != nil {
-		return nil, err
-	}
-	req.Encoding = encoding.String()
-
-	s = r.URL.Query().Get(urlParamIndexPageSize)
-	indexPageSize, err := strconv.ParseInt(s, 10, 32)
-	if err != nil {
-		return nil, fmt.Errorf("invalid indexPageSize %s: %w", s, err)
-	}
-	req.IndexPageSize = uint32(indexPageSize)
-
-	s = r.URL.Query().Get(urlParamTotalRecords)
-	totalRecords, err := strconv.ParseInt(s, 10, 32)
-	if err != nil {
-		return nil, fmt.Errorf("invalid totalRecords %s: %w", s, err)
-	}
-	if totalRecords <= 0 {
-		return nil, fmt.Errorf("totalRecords must be greater than 0. received %d", totalRecords)
-	}
-	req.TotalRecords = uint32(totalRecords)
-
-	// Data encoding can be blank for some block formats, therefore
-	// no validation on the param here.  Eventually we may be able
-	// to remove this parameter entirely.
-	dataEncoding := r.URL.Query().Get(urlParamDataEncoding)
-	req.DataEncoding = dataEncoding
-
-	version := r.URL.Query().Get(urlParamVersion)
-	if version == "" {
-		return nil, errors.New("version required")
-	}
-	req.Version = version
-
-	s = r.URL.Query().Get(urlParamSize)
-	size, err := strconv.ParseUint(s, 10, 64)
-	if err != nil {
-		return nil, fmt.Errorf("invalid size %s: %w", s, err)
-	}
-	req.Size_ = size
-
-	// Footer size can be 0 for some blocks, just ensure we
-	// get a valid integer.
-	f := r.URL.Query().Get(urlParamFooterSize)
-	footerSize, err := strconv.ParseUint(f, 10, 32)
-	if err != nil {
-		return nil, fmt.Errorf("invalid footerSize %s: %w", f, err)
-	}
-	req.FooterSize = uint32(footerSize)
-
-	s = r.URL.Query().Get(urlParamDedicatedColumns)
-	if s != "" {
-		var dedicatedColumns []*tempopb.DedicatedColumn
-		err = json.Unmarshal([]byte(s), &dedicatedColumns)
-		if err != nil {
-			return nil, fmt.Errorf("invalid dedicatedColumns '%s': %w", s, err)
-		}
-		req.DedicatedColumns = dedicatedColumns
-	}
-
 	return req, nil
 }
 
@@ -868,106 +530,6 @@ func BuildSearchRequest(req *http.Request, searchReq *tempopb.SearchRequest) (*h
 
 		q.Set(urlParamTags, builder.String())
 	}
-
-	req.URL.RawQuery = q.Encode()
-
-	return req, nil
-}
-
-func BuildSearchTagRequest(req *http.Request, searchReq *tempopb.SearchTagsRequest) (*http.Request, error) {
-	if req == nil {
-		req = &http.Request{
-			URL: &url.URL{},
-		}
-	}
-
-	if searchReq == nil {
-		return req, nil
-	}
-
-	q := req.URL.Query()
-	q.Set(urlParamStart, strconv.FormatUint(uint64(searchReq.Start), 10))
-	q.Set(urlParamEnd, strconv.FormatUint(uint64(searchReq.End), 10))
-	q.Set(urlParamScope, searchReq.Scope)
-
-	req.URL.RawQuery = q.Encode()
-
-	return req, nil
-}
-
-func BuildSearchTagsBlockRequest(req *http.Request, searchReq *tempopb.SearchTagsBlockRequest) (*http.Request, error) {
-	if req == nil {
-		req = &http.Request{
-			URL: &url.URL{},
-		}
-	}
-
-	req, err := BuildSearchTagRequest(req, searchReq.SearchReq)
-	if err != nil {
-		return nil, err
-	}
-
-	q := req.URL.Query()
-	q.Set(urlParamSize, strconv.FormatUint(searchReq.Size_, 10))
-	q.Set(urlParamBlockID, searchReq.BlockID)
-	q.Set(urlParamStartPage, strconv.FormatUint(uint64(searchReq.StartPage), 10))
-	q.Set(urlParamPagesToSearch, strconv.FormatUint(uint64(searchReq.PagesToSearch), 10))
-	q.Set(urlParamEncoding, searchReq.Encoding)
-	q.Set(urlParamIndexPageSize, strconv.FormatUint(uint64(searchReq.IndexPageSize), 10))
-	q.Set(urlParamTotalRecords, strconv.FormatUint(uint64(searchReq.TotalRecords), 10))
-	q.Set(urlParamDataEncoding, searchReq.DataEncoding)
-	q.Set(urlParamVersion, searchReq.Version)
-	q.Set(urlParamFooterSize, strconv.FormatUint(uint64(searchReq.FooterSize), 10))
-
-	req.URL.RawQuery = q.Encode()
-
-	return req, nil
-}
-
-func BuildSearchTagValuesRequest(req *http.Request, searchReq *tempopb.SearchTagValuesRequest) (*http.Request, error) {
-	if req == nil {
-		req = &http.Request{
-			URL: &url.URL{},
-		}
-	}
-
-	if searchReq == nil {
-		return req, nil
-	}
-
-	q := req.URL.Query()
-	q.Set(urlParamStart, strconv.FormatUint(uint64(searchReq.Start), 10))
-	q.Set(urlParamEnd, strconv.FormatUint(uint64(searchReq.End), 10))
-	q.Set(urlParamQuery, searchReq.Query)
-
-	req.URL.RawQuery = q.Encode()
-
-	return req, nil
-}
-
-func BuildSearchTagValuesBlockRequest(req *http.Request, searchReq *tempopb.SearchTagValuesBlockRequest) (*http.Request, error) {
-	if req == nil {
-		req = &http.Request{
-			URL: &url.URL{},
-		}
-	}
-
-	req, err := BuildSearchTagValuesRequest(req, searchReq.SearchReq)
-	if err != nil {
-		return nil, err
-	}
-
-	q := req.URL.Query()
-	q.Set(urlParamSize, strconv.FormatUint(searchReq.Size_, 10))
-	q.Set(urlParamBlockID, searchReq.BlockID)
-	q.Set(urlParamStartPage, strconv.FormatUint(uint64(searchReq.StartPage), 10))
-	q.Set(urlParamPagesToSearch, strconv.FormatUint(uint64(searchReq.PagesToSearch), 10))
-	q.Set(urlParamEncoding, searchReq.Encoding)
-	q.Set(urlParamIndexPageSize, strconv.FormatUint(uint64(searchReq.IndexPageSize), 10))
-	q.Set(urlParamTotalRecords, strconv.FormatUint(uint64(searchReq.TotalRecords), 10))
-	q.Set(urlParamDataEncoding, searchReq.DataEncoding)
-	q.Set(urlParamVersion, searchReq.Version)
-	q.Set(urlParamFooterSize, strconv.FormatUint(uint64(searchReq.FooterSize), 10))
 
 	req.URL.RawQuery = q.Encode()
